+++ pallets/transfer-allowlist/Cargo.toml
@@ -20,8 +20,6 @@ cfg-traits = { path = "../../libs/traits", default-features = false }
 
 cfg-types = { path = "../../libs/types", default-features = false }
 
-cfg-test-utils = { path = "../../libs/test-utils", default-features = false }
-
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", default-features = false, optional = true, branch = "polkadot-v0.9.37" }
 frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.37" }
 frame-system = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.37" }
