+++ pallets/transfer-allowlist/src/lib.rs
@@ -22,7 +22,7 @@
 //! - the block range specified in the allowance
 
 #[cfg(test)]
-mod mock;
+pub(crate) mod mock;
 
 #[cfg(test)]
 mod tests;
