+++ pallets/transfer-allowlist/src/tests.rs
@@ -1,4 +1,3 @@
-use cfg_test_utils::system::time_travel::advance_n_blocks;
 use frame_support::{assert_err, assert_noop, assert_ok};
 
 use super::*;
